Arranging Files according to the convention
