+++ utils/saveImage.go
@@ -1,0 +1,17 @@
+package utils
+
+import (
+	"image"
+	"image/png"
+	"os"
+)
+
+func SaveImageToPng(filename string, img image.Image) error {
+	f, err := os.OpenFile(filename, os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+
+	return png.Encode(f, img)
+}
